Add 'East Wing' to Hub Base template
07.28.2023

- fix: revert back baseMapPixels to int[][]
- change: add new chinese and uk names
- change: add 'East Wing' with new buildings in
          settlement_hub_base.xml

+++ mars-sim-mapdata/src/main/java/org/mars/sim/mapdata/IntegerMapData.java
@@ -50,11 +50,11 @@ import com.jogamp.opencl.CLProgram;
  	private final double RHO_DEFAULT;
 
  	// Data members.
- 	private short[][] baseMapPixels = null;
+ 	private int[][] baseMapPixels = null;
  	// # of pixels in the width of the map image
-	private short pixelWidth;
+	private int pixelWidth;
  	// # of pixels in the height of the map image
-	private short pixelHeight;
+	private int pixelHeight;
 	// height pixels divided by pi (equals to pixelHeight / Math.PI)
 	private double rho;
 	
@@ -189,13 +189,13 @@ import com.jogamp.opencl.CLProgram;
  	 * @return
  	 * @throws IOException
  	 */
- 	private short[][] loadMapData(String imageName) throws IOException {
+ 	private int[][] loadMapData(String imageName) throws IOException {
 
  		File imageFile = FileLocator.locateFile(imageName);
  		
  		BufferedImage cylindricalMapImage = null;
  		
- 		short[][] mapPixels = null;
+ 		int[][] mapPixels = null;
  		
 		try {
 			cylindricalMapImage = ImageIO.read(imageFile);
@@ -207,12 +207,12 @@ import com.jogamp.opencl.CLProgram;
 			final byte[] pixels = ((DataBufferByte) cylindricalMapImage.getRaster().getDataBuffer()).getData();
 //	 		int[] srcPixels = ((DataBufferInt)cylindricalMapImage.getRaster().getDataBuffer()).getData();
 
-	 		pixelWidth = (short)cylindricalMapImage.getWidth();
-	 		pixelHeight = (short)cylindricalMapImage.getHeight();
+	 		pixelWidth = cylindricalMapImage.getWidth();
+	 		pixelHeight = cylindricalMapImage.getHeight();
 	 		
 	 		final boolean hasAlphaChannel = cylindricalMapImage.getAlphaRaster() != null;
 		
-	 		mapPixels = new short[pixelHeight][pixelWidth];
+	 		mapPixels = new int[pixelHeight][pixelWidth];
 	 		
 	 		if (hasAlphaChannel) {
 	 			final int pixelLength = 4;
@@ -242,7 +242,7 @@ import com.jogamp.opencl.CLProgram;
 //	 				do ((int) pixels[pixel + pixel_offset + 1] & 0xff); // green; 
 //	 				and merge the two loops into one. – Tomáš Zato Mar 23 '15 at 23:02
 	 						
-	 				mapPixels[row][col] = (short)argb;
+	 				mapPixels[row][col] = argb;
 	 				col++;
 	 				if (col == pixelWidth) {
 	 					col = 0;
@@ -261,7 +261,7 @@ import com.jogamp.opencl.CLProgram;
 	 				argb += (((int) pixels[pixel + 1] & 0xff) << 8); // green
 	 				argb += (((int) pixels[pixel + 2] & 0xff) << 16); // red
 	 				
-	 				mapPixels[row][col] = (short)argb;
+	 				mapPixels[row][col] = argb;
 	 				col++;
 	 				if (col == pixelWidth) {
 	 					col = 0;
@@ -558,7 +558,7 @@ import com.jogamp.opencl.CLProgram;
 // 		return cylindricalMapImage;
 // 	}
  	
- 	public short[][] getPixels() {
+ 	public int[][] getPixels() {
  		return baseMapPixels;
  	}
